Make ShaderEffect an enum class
The effect is no longer compared by its position in the enumeration and every
user passes the typed value around, so it can be a scoped enumeration.
QMetaEnum still stores and reads it by name; valueToKey() now receives the
value through an explicit cast.

+++ apps/quickviewer/src/models/shadermanager.h
@@ -34,7 +34,7 @@ public:
     static QString shaderEffectToString(qvEnums::ShaderEffect effect)
     {
         QMetaEnum metaEnum = QMetaEnum::fromType<qvEnums::ShaderEffect>();
-        return metaEnum.valueToKey(effect);
+        return metaEnum.valueToKey(static_cast<int>(effect));
     }
     static qvEnums::ShaderEffect stringToShaderEffect(QString effect)
     {
